requirements.txt
# These requirements are for local development and testing of the module

# python packaging
twine==4.0.2

# django stuff
Django==5.0.1
djangorestframework==3.14.0
pytz==2023.3

# factories
factory-boy==3.2.1

# tests
pytest==7.3.2
pytest-django==4.5.2
flake8==6.0.0

# payment processing
dj-stripe==2.8.1
